# apps/core-server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/commerce-api/orders/table": {
      "get": {
        "operationId": "OrdersGatewayController_getOrderTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Orders"
        ]
      }
    },
...
    "/commerce-api/orders/{id}": {
      "get": {
        "operationId": "OrdersGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Orders"
        ]
      }
    },
    "/commerce-api/orders/{id}/status": {
      "post": {
        "operationId": "OrdersGatewayController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Orders"
        ]
      }
    },
...
    "/commerce-api/purchase-orders/table": {
      "get": {
        "operationId": "PurchaseOrdersGatewayController_getTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
    "/commerce-api/purchase-orders": {
      "post": {
        "operationId": "PurchaseOrdersGatewayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
    "/commerce-api/purchase-orders/{id}/pdf": {
      "get": {
        "operationId": "PurchaseOrdersGatewayController_downloadPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
    "/commerce-api/purchase-orders/{id}": {
      "get": {
        "operationId": "PurchaseOrdersGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      },
      "patch": {
        "operationId": "PurchaseOrdersGatewayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePurchaseOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      },
      "delete": {
        "operationId": "PurchaseOrdersGatewayController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
    "/commerce-api/purchase-orders/{id}/status": {
      "patch": {
        "operationId": "PurchaseOrdersGatewayController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
    "/commerce-api/purchase-orders/{id}/send-email": {
      "post": {
        "operationId": "PurchaseOrdersGatewayController_sendEmail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPurchaseOrderEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Purchase Orders"
        ]
      }
    },
...
    "/commerce-api/suppliers/table": {
      "get": {
        "operationId": "SuppliersGatewayController_getSupplierTable",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers/select": {
      "get": {
        "operationId": "SuppliersGatewayController_select",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter by label",
            "schema": {
              "example": "united",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of results",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "values",
            "required": false,
            "in": "query",
            "description": "Comma-separated values to fetch specific options",
            "schema": {
              "example": "1,2,3",
              "type": "string"
            }
          },
          {
            "name": "excludeIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated IDs to exclude from results (already selected)",
            "schema": {
              "example": "5,10",
              "type": "string"
            }
          },
          {
            "name": "valueKey",
            "required": false,
            "in": "query",
            "description": "Column name for option value",
            "schema": {
              "default": "id",
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "labelKey",
            "required": false,
            "in": "query",
            "description": "Column name for option label",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "descriptionKey",
            "required": false,
            "in": "query",
            "description": "Column name for option description",
            "schema": {
              "example": "description",
              "type": "string"
            }
          },
          {
            "name": "additionalKeys",
            "required": false,
            "in": "query",
            "description": "Comma-separated column names to include in option.additionals",
            "schema": {
              "example": "locationName,availableQuantity",
              "type": "string"
            }
          },
          {
            "name": "groupIdKey",
            "required": false,
            "in": "query",
            "description": "Column name for group ID",
            "schema": {
              "example": "regionId",
              "type": "string"
            }
          },
          {
            "name": "orderByKey",
            "required": false,
            "in": "query",
            "description": "Column name to sort options by",
            "schema": {
              "default": "name",
              "example": "name",
              "type": "string"
            }
          },
          {
            "name": "orderDirection",
            "required": false,
            "in": "query",
            "description": "Sort direction for options",
            "schema": {
              "default": "asc",
              "example": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers": {
      "post": {
        "operationId": "SuppliersGatewayController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers/items/price": {
      "get": {
        "operationId": "SuppliersGatewayController_getItemPrice",
        "parameters": [
          {
            "name": "supplierId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "inventoryItemId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers/{id}": {
      "get": {
        "operationId": "SuppliersGatewayController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      },
      "patch": {
        "operationId": "SuppliersGatewayController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      },
      "delete": {
        "operationId": "SuppliersGatewayController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers/{id}/items": {
      "post": {
        "operationId": "SuppliersGatewayController_linkItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkSupplierItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
    "/commerce-api/suppliers/items/{itemId}": {
      "delete": {
        "operationId": "SuppliersGatewayController_unlinkItem",
        "parameters": [
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Commerce - Suppliers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vritti API Nexus",
    "description": "Internal API for Vritti Nexus Platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://local.vrittiai.com:3001",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT access token"
      }
    },
    "schemas": {
...
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Target order status",
            "example": "ACCEPTED",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "PREPARING",
              "READY",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "cancellationReason": {
            "type": "string",
            "description": "Reason for cancellation (required when cancelling)"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreatePurchaseOrderItemDto": {
        "type": "object",
        "properties": {
          "inventoryItemId": {
            "type": "string",
            "description": "Inventory item ID"
          },
          "orderedQuantity": {
            "type": "number",
            "description": "Ordered quantity",
            "example": 100
          }
        },
        "required": [
          "inventoryItemId",
          "orderedQuantity"
        ]
      },
      "CreatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "supplierId": {
            "type": "string",
            "description": "Supplier ID"
          },
          "poNumber": {
            "type": "string",
            "description": "PO number (auto-generated if omitted)"
          },
          "orderDate": {
            "type": "string",
            "description": "Order date (ISO string)",
            "example": "2026-04-10"
          },
          "expectedDate": {
            "type": "string",
            "description": "Expected delivery date (ISO string)"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          },
          "items": {
            "description": "Line items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreatePurchaseOrderItemDto"
            }
          }
        },
        "required": [
          "supplierId",
          "orderDate"
        ]
      },
      "UpdatePurchaseOrderItemDto": {
        "type": "object",
        "properties": {
          "inventoryItemId": {
            "type": "string",
            "description": "Inventory item ID"
          },
          "orderedQuantity": {
            "type": "number",
            "description": "Ordered quantity",
            "example": 100
          },
          "unitPrice": {
            "type": "object",
            "description": "Unit price",
            "example": 99.99
          }
        }
      },
      "UpdatePurchaseOrderDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "PO status",
            "enum": [
              "DRAFT",
              "SENT",
              "PARTIALLY_RECEIVED",
              "RECEIVED",
              "CANCELLED"
            ]
          },
          "orderDate": {
            "type": "string",
            "description": "Order date (ISO string)"
          },
          "expectedDate": {
            "type": "string",
            "description": "Expected delivery date (ISO string)"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          },
          "totalAmount": {
            "type": "object",
            "description": "Total amount (null to clear)",
            "nullable": true
          },
          "items": {
            "description": "Replace all line items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdatePurchaseOrderItemDto"
            }
          }
        }
      },
      "SendPurchaseOrderEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Override recipient email; defaults to supplier email"
          }
        }
      },
...
      "CreateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Supplier name",
            "example": "Acme Foods Pvt Ltd"
          },
          "code": {
            "type": "string",
            "description": "Unique supplier code",
            "example": "SUP-001"
          },
          "contactName": {
            "type": "string",
            "description": "Primary contact person name"
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "email": {
            "type": "string",
            "description": "Email address"
          },
          "address": {
            "type": "string",
            "description": "Mailing address"
          },
          "gstin": {
            "type": "string",
            "description": "GST identification number",
            "example": "22AAAAA0000A1Z5"
          },
          "paymentTerms": {
            "type": "string",
            "description": "Payment terms",
            "example": "Net 30"
          },
          "leadTimeDays": {
            "type": "number",
            "description": "Default lead time in days",
            "example": 7
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "UpdateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated supplier name"
          },
          "code": {
            "type": "string",
            "description": "Updated supplier code"
          },
          "contactName": {
            "type": "string",
            "description": "Updated contact person name"
          },
          "phone": {
            "type": "string",
            "description": "Updated phone number"
          },
          "email": {
            "type": "string",
            "description": "Updated email address"
          },
          "address": {
            "type": "string",
            "description": "Updated mailing address"
          },
          "gstin": {
            "type": "string",
            "description": "Updated GSTIN"
          },
          "paymentTerms": {
            "type": "string",
            "description": "Updated payment terms"
          },
          "leadTimeDays": {
            "type": "number",
            "description": "Updated lead time in days"
          },
          "notes": {
            "type": "string",
            "description": "Updated notes"
          }
        }
      },
      "LinkSupplierItemDto": {
        "type": "object",
        "properties": {
          "inventoryItemId": {
            "type": "string",
            "description": "Inventory item ID to link"
          },
          "supplierCode": {
            "type": "string",
            "description": "Supplier-specific item code"
          },
          "unitPrice": {
            "type": "number",
            "description": "Unit price from this supplier",
            "example": 12.5
          },
          "uomId": {
            "type": "string",
            "description": "UOM ID for pricing"
          },
          "minOrderQuantity": {
            "type": "number",
            "description": "Minimum order quantity",
            "example": 10
          },
          "leadTimeDays": {
            "type": "number",
            "description": "Lead time in days for this item",
            "example": 5
          },
          "isPreferred": {
            "type": "boolean",
            "description": "Whether this is the preferred supplier for this item",
            "default": false
          }
        },
        "required": [
          "inventoryItemId",
          "uomId"
        ]
      },
...
    }
  }
}